setup updated search component
