+++ client/src/components/Header/Header.scss
@@ -4,7 +4,7 @@
 .header {
     text-align: center;
     position: relative;
-    cursor: none;
+    // cursor: none;
 
     .logo {
         $logo_size: 3px;
