+++ client/src/components/Search/Search.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/colors';
-@import '../../../styles/sizes';
+@import '../../styles/colors';
+@import '../../styles/sizes';
 
 .search-button {
     cursor: pointer;
@@ -18,13 +18,9 @@
     left: 0px;
     width: 100%;
     height: 100%;
+    padding: 100px 0px;
     background: rgba(0,0,8,0.9);
     display: flex;
     justify-content: center;
-    align-items: center;
     z-index: 100;
-}
-
-.search-content {
-    width: 600px;
 }
